+++ include/fluxgl/ecs/components/mesh_renderer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <fluxgl/graphics/material.h>
-#include <fluxgl/graphics/mesh.h>
+#include <fluxgl/assets/resource_manager.h>
 
 #include <json/json.h>
 #include <fluxgl/ecs/registry.h>
@@ -10,13 +10,13 @@
 
 namespace fluxgl {
     struct MeshRenderer {
-        // static constexpr const char* TypeName = "MeshRenderer";
-        // static void Create(fluxgl::Registry& registry, fluxgl::EntityID entityID, const Json::Value& data) {
-        //     MeshRenderer renderer;
-        //     registry.addComponent<MeshRenderer>(entityID, renderer);
-        // }
+        static constexpr const char* TypeName = "MeshRenderer";
+        static void Create(fluxgl::Registry& registry, fluxgl::EntityID entityID, const Json::Value& data) {
+            MeshRenderer renderer;
+            registry.addComponent<MeshRenderer>(entityID, renderer);
+        }
 
         Material material;
-        Mesh mesh;
+        MeshHandle mesh;
     };
 }