+++ include/fluxgl/ecs/components/sprite.h
@@ -1,22 +1,20 @@
 #pragma once
 
-#include <fluxgl/graphics/texture.h>
-#include <fluxgl/graphics/shader.h>
-
 #include <json/json.h>
 #include <fluxgl/ecs/registry.h>
 #include <fluxgl/ecs/entity.h>
+#include <fluxgl/assets/resource_manager.h>
 
 namespace fluxgl {
     struct Sprite {
-        // static constexpr const char* TypeName = "Sprite";
-        // static void Create(fluxgl::Registry& registry, fluxgl::EntityID entityID, const Json::Value& data) {
-        //     Sprite sprite;
-        //     registry.addComponent<Sprite>(entityID, sprite);
-        // }
+        static constexpr const char* TypeName = "Sprite";
+        static void Create(fluxgl::Registry& registry, fluxgl::EntityID entityID, const Json::Value& data) {
+            Sprite sprite;
+            registry.addComponent<Sprite>(entityID, sprite);
+        }
 
-        Texture texture;
-        Shader shader;
+        TextureHandle texture;
+        ShaderHandle shader;
 
         glm::vec3 color = {1.0f, 1.0f, 1.0f};
         glm::vec2 size = {1.0f, 1.0f};