+++ include/fluxgl/graphics/material.h
@@ -2,21 +2,22 @@
 
 #include <glm/glm.hpp>
 #include <vector>
+#include <fluxgl/assets/resource_manager.h>
 #include "shader.h"
 #include "texture.h"
 
 namespace fluxgl {
     struct Material {
-        Shader shader;
+        ShaderHandle shader;
 
         glm::vec3 albedoColor = glm::vec3(1.0f);
         glm::vec3 specularColor = glm::vec3(1.0f);
         glm::vec3 emissionColor = glm::vec3(0.0f);
         float shininess = 32.0f;
 
-        std::vector<Texture> albedoTextures;
-        Texture normalMap;
-        Texture specularMap;
-        Texture emissionMap;
+        std::vector<TextureHandle> albedoTextures;
+        TextureHandle normalMap;
+        TextureHandle specularMap;
+        TextureHandle emissionMap;
     };
 }